test(sccm): restrict raw management point reducer
